refs #963
fix gincore import

+++ manage/modules/import/import_gincore_items.php
@@ -157,9 +157,15 @@ class import_gincore_items extends abstract_import_handler
         return $data;
     }
 
+    /**
+     * @param $id
+     * @param $value
+     */
     private function setCategory($id, $value)
     {
-        db()->query('DELETE FROM {category_goods} WHERE goods_id=?i', array($id));
-        db()->query('INSERT INTO {category_goods} (goods_id, category_id) VALUES (?i, ?i)', array($id, $value));
+        if (!empty($value)) {
+            db()->query('DELETE FROM {category_goods} WHERE goods_id=?i', array($id));
+            db()->query('INSERT INTO {category_goods} (goods_id, category_id) VALUES (?i, ?i)', array($id, $value));
+        }
     }
 }
